fix: extract community builds card to client component
Move BuildCard rendering to CommunityBuildsList client component to
avoid passing functions as props to client components, which fails
during server rendering.
